@@ -3,20 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.5",
-    "@testing-library/user-event": "^12.7.3",
+    "@testing-library/jest-dom": "^6.4.6",
+    "@testing-library/react": "^16.0.0",
+    "@testing-library/user-event": "^14.5.2",
     "@tsparticles/react": "^3.0.0",
     "clarifai": "^2.9.1",
-    "postcss": "^8.3.5",
-    "postcss-safe-parser": "^5.0.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "postcss": "^8.4.38",
+    "postcss-safe-parser": "^7.0.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-parallax-tilt": "^1.7.228",
-    "react-scripts": "4.0.3",
+    "react-scripts": "5.0.1",
     "tachyons": "^4.12.0",
     "tsparticles": "^3.4.0",
-    "web-vitals": "^1.1.0"
+    "web-vitals": "^4.2.0"
   },
   "scripts": {
     "start": "set PORT=3001 && set NODE_OPTIONS=--openssl-legacy-provider && react-scripts start",
