@@ -1,115 +1,53 @@
 const express = require("express");
-const bodyParser = require("body-parser"); // latest version of exressJS now comes with Body-Parser!
+const bodyParser = require("body-parser");
 const bcrypt = require("bcrypt-nodejs");
 const cors = require("cors");
 const knex = require("knex");
 
+const register = require("./controllers/register");
+const signin = require("./controllers/signin");
+const profile = require("./controllers/profile");
+const image = require("./controllers/image");
+
+process.env.NODE_TLS_REJECT_UNAUTHORIZED = 0;
+
 const db = knex({
-  // Enter your own database information here based on what you created
   client: "pg",
   connection: {
-    host: "127.0.0.1",
-    user: "aneagoie",
-    password: "",
-    database: "smart-brain",
+    connectionString: process.env.DATABASE_URL,
+    ssl: {
+      rejectUnauthorized: false,
+    },
   },
 });
 
 const app = express();
 
+app.use(bodyParser.json());
 app.use(cors());
-app.use(express.json()); // latest version of exressJS now comes with Body-Parser!
 
-// Test only - when you have a database variable you want to use
-// app.get('/', (req, res)=> {
-//   res.send(database.users);
-// })
-
-app.post("/signin", (req, res) => {
-  db.select("email", "hash")
-    .from("login")
-    .where("email", "=", req.body.email)
-    .then((data) => {
-      const isValid = bcrypt.compareSync(req.body.password, data[0].hash);
-      if (isValid) {
-        return db
-          .select("*")
-          .from("users")
-          .where("email", "=", req.body.email)
-          .then((user) => {
-            res.json(user[0]);
-          })
-          .catch((err) => res.status(400).json("unable to get user"));
-      } else {
-        res.status(400).json("wrong credentials");
-      }
-    })
-    .catch((err) => res.status(400).json("wrong credentials"));
+app.get("/", (req, res) => {
+  res.send("it is working!");
 });
 
+app.post("/signin", signin.handleSignin(db, bcrypt));
+
 app.post("/register", (req, res) => {
-  const { email, name, password } = req.body;
-  const hash = bcrypt.hashSync(password);
-  db.transaction((trx) => {
-    trx
-      .insert({
-        hash: hash,
-        email: email,
-      })
-      .into("login")
-      .returning("email")
-      .then((loginEmail) => {
-        return trx("users")
-          .returning("*")
-          .insert({
-            // If you are using knex.js version 1.0.0 or higher this now returns an array of objects. Therefore, the code goes from:
-            // loginEmail[0] --> this used to return the email
-            // TO
-            // loginEmail[0].email --> this now returns the email
-            email: loginEmail[0].email,
-            name: name,
-            joined: new Date(),
-          })
-          .then((user) => {
-            res.json(user[0]);
-          });
-      })
-      .then(trx.commit)
-      .catch(trx.rollback);
-  }).catch((err) => res.status(400).json("unable to register"));
+  register.handleRegister(req, res, db, bcrypt);
 });
 
 app.get("/profile/:id", (req, res) => {
-  const { id } = req.params;
-  db.select("*")
-    .from("users")
-    .where({ id })
-    .then((user) => {
-      if (user.length) {
-        res.json(user[0]);
-      } else {
-        res.status(400).json("Not found");
-      }
-    })
-    .catch((err) => res.status(400).json("error getting user"));
+  profile.handleProfileGet(req, res, db);
 });
 
 app.put("/image", (req, res) => {
-  const { id } = req.body;
-  db("users")
-    .where("id", "=", id)
-    .increment("entries", 1)
-    .returning("entries")
-    .then((entries) => {
-      // If you are using knex.js version 1.0.0 or higher this now returns an array of objects. Therefore, the code goes from:
-      // entries[0] --> this used to return the entries
-      // TO
-      // entries[0].entries --> this now returns the entries
-      res.json(entries[0].entries);
-    })
-    .catch((err) => res.status(400).json("unable to get entries"));
+  image.handleImage(req, res, db);
 });
 
-app.listen(3000, () => {
-  console.log("app is running on port 3000");
+app.post("/imageurl", (req, res) => {
+  image.handleApiCall(req, res);
+});
+
+app.listen(process.env.PORT || 3000, () => {
+  console.log(`app is running on port ${process.env.PORT}`);
 });
